update clhm to 1.3 patch by Vijay; reviewed by jbellis for CASSANDRA-4150

+++ src/java/org/apache/cassandra/cache/ICache.java
@@ -30,9 +30,9 @@ import java.util.Set;
  */
 public interface ICache<K, V>
 {
-    public int capacity();
+    public long capacity();
 
-    public void setCapacity(int capacity);
+    public void setCapacity(long capacity);
 
     public void put(K key, V value);
 
@@ -46,7 +46,7 @@ public interface ICache<K, V>
 
     public int size();
 
-    public int weightedSize();
+    public long weightedSize();
 
     public void clear();
